requirements.txt
# Multi-Tool AI Agent dependencies
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
duckduckgo-search>=6.3.0
pydantic>=2.0.0
